Modify unit test

+++ feature/characters/src/test/java/io/fajarca/characters/util/TestUtil.kt
@@ -4,7 +4,7 @@ import io.fajarca.core.database.CharacterEntity
 
 object TestUtil {
 
-    fun generateDummyCharacter(id : Long, title : String = "Title", imageUrl : String = "http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available/portrait_uncanny.jpg" ) : CharacterEntity {
+    fun generateDummyCharacter(id : Int, title : String = "Title", imageUrl : String = "http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available/portrait_uncanny.jpg" ) : CharacterEntity {
         return CharacterEntity(
             id = id,
             title = title,
@@ -14,7 +14,7 @@ object TestUtil {
     fun generateDummyCharacters(numOfIteration : Int): List<CharacterEntity> {
         val characters = mutableListOf<CharacterEntity>()
         for (i in 1 .. numOfIteration) {
-            characters.add(generateDummyCharacter(i.toLong()))
+            characters.add(generateDummyCharacter(i))
         }
         return characters
     }
